refactor help-tips
